+++ .test_durations
@@ -1640,8 +1640,8 @@
     "tests/valuation/test_interface.py::test_data_banzhaf_valuation[2]": 1.2780167490000025,
     "tests/valuation/test_interface.py::test_data_beta_shapley_valuation[1]": 4.139234666999997,
     "tests/valuation/test_interface.py::test_data_beta_shapley_valuation[2]": 3.603092916999998,
-    "tests/valuation/test_interface.py::test_data_shapley_valuation[1]": 0.27120083299999465,
-    "tests/valuation/test_interface.py::test_data_shapley_valuation[2]": 0.15037520699999618,
+    "tests/valuation/test_interface.py::test_shapley_valuation[1]": 0.27120083299999465,
+    "tests/valuation/test_interface.py::test_shapley_valuation[2]": 0.15037520699999618,
     "tests/valuation/test_interface.py::test_data_utility_learning[1]": 0.026216332999993597,
     "tests/valuation/test_interface.py::test_data_utility_learning[2]": 0.06457645800000478,
     "tests/valuation/test_interface.py::test_delta_shapley_valuation[1]": 3.562169998999977,
