+++ src/pydvl/valuation/methods/__init__.py
@@ -2,7 +2,6 @@ from .beta_shapley import *
 from .classwise_shapley import *
 from .data_banzhaf import *
 from .data_oob import *
-from .data_shapley import *
 from .delta_shapley import *
 from .gt_shapley import *
 from .knn_shapley import *
@@ -10,3 +9,4 @@ from .least_core import *
 from .loo import *
 from .owen_shapley import *
 from .semivalue import *
+from .shapley import *
